+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - 新增选项 `only.start.page`
 - 增加选项 `title.in.journal`
 - 增加选项 `space.before.pages`
+- 增加选项 `abbreviate.journal`
 
 ## [v2.0] - 2020-03-04
 ### Changed